Beep when attempting to open a non-openable file
When you first start working on a local branch, and you haven't yet pushed it up
to GitHub, we can't show you to the file on GitHub (because the branch doesn't
exist on GitHub). This is a fairly common scenario. Prior to this commit,
we would just write a warning to the console in this scenario. Since the user
usually has the console closed, this can lead the user to think that
to-the-hubs just isn't working. :grimacing:

Until we have a better (more visual) way to alert the user in these scenarios,
let's use a system beep to provide a more obvious indication that something is
amiss.