Default to 'master' for local branches
This adds a check to see if the current branch is local or has been pushed. If the branch is local, it will preform the 'open', 'blame', and 'history' commands by opening the master branch on github. If it is available on the remote it will preform the operation using the current branch name.